msm: mdss: save state of vsync handler
When going into low power modes, vsync handlers can be forcefully shut
down in order to disable clocks. If vsync handler was expected to be
enabled it would lose that state. Instead maintain the state of the
vsync and re-enable on next panel on.

Bug: 62375787
Git-commit: dea49443c3e803158fb213a501452431b76fce3e
Git-repo: https://android.googlesource.com/kernel/common
Change-Id: Ic4e7d666f17852e5d6c60979151e27ab3fbf952d
Signed-off-by: Adrian Salido <salidoa@google.com>
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>