Revert "msm: arm64: fix compiler warnings due to improper braces or indentation"
This reverts commit d97e82ee1caca6ac0bdfab3cdf96f223e16419f5.

This is incorrect because driver mandates for conn.dl and conn.ul to be enabled
together, hence conn.ul will always be free from memory allocation if conn.dl is
disabled and vice versa.

The one squashed from 'compiler-fixes' branch is the proper fix, which frees
conn.ul only if conn.dl is enabled.

Signed-off-by: Albert I <kras@raphielgang.org>